fix life time issue  in generate token

+++ api-service/app/Services/Auth/AccessTokenService.php
@@ -16,7 +16,7 @@ class AccessTokenService
 
         $tokenObject = $user->createToken(
             name: $requestDTO->getTokenName(),
-            expiresAt: $requestDTO->getExpirationDate() ?? now()->addMinutes(config('sanctum.expiration'))
+            expiresAt: $requestDTO->getExpirationDate() ?? now()->addMinutes((int) config('sanctum.expiration'))
         );
         // Update token details with IP and user agent
         $tokenObject->accessToken->ip = $requestDTO->getIpAddress();
